# demart-api-contract.yaml
openapi: 3.0.3
info:
  title: DeMart API
  version: 1.0.0
  description: REST API contract for the DeMart e-commerce application.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: Get products
      operationId: getProducts
      responses:
        '200':
          description: Products returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - products
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
        '500':
          description: Failed to retrieve products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Products]
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Product already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/orders:
    get:
      tags: [Orders]
      summary: Get the authenticated user's orders
      operationId: getOrders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - orders
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Orders]
      summary: Create an order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Invalid order request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Order or inventory conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: Get all users
      operationId: getAdminUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/orders:
    get:
      tags: [Admin]
      summary: Get all orders
      operationId: getAdminOrders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Orders returned successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - orders
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/stores:
    get:
      tags: [Stores]
      summary: Get stores for a location
      operationId: getStoresByLocation
      parameters:
        - name: locationId
          in: query
          required: true
          description: Location ID.
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Stores returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [stores]
                properties:
                  stores:
                    type: array
                    items:
                      $ref: '#/components/schemas/Store'
        '400':
          description: Location ID is missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to load stores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - street
        - postalCode
        - city
        - country
        - role
      properties:
        id:
          type: integer
          example: 101
        firstName:
          type: string
          example: Vinod
        lastName:
          type: string
          example: Arokiaraj
        email:
          type: string
          format: email
          example: customer@example.com
        phone:
          type: string
          example: '+4915123456789'
        street:
          type: string
          example: Main Street 10
        addressLine2:
          type: string
          nullable: true
          example: Apartment 4B
        landmark:
          type: string
          nullable: true
          example: Near Central Station
        postalCode:
          type: string
          example: '60311'
        city:
          type: string
          example: Frankfurt
        country:
          type: string
          example: Germany
        role:
          type: string
          enum: [CUSTOMER, ADMIN]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProductRequest:
      type: object
      required:
        - name
        - category
        - price
        - stockQuantity
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
          minimum: 0
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          nullable: true
        stockQuantity:
          type: integer
          minimum: 0
        imageUrl:
          type: string
          nullable: true
...
    ProductResponse:
      type: object
      required:
        - message
        - product
      properties:
        message:
          type: string
        product:
          $ref: '#/components/schemas/Product'
    Product:
      type: object
      required:
        - id
        - name
        - category
        - price
        - rating
        - stockQuantity
        - imageUrl
        - isActive
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
          minimum: 0
          example: 49.99
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          example: 4.7
        stockQuantity:
          type: integer
          minimum: 0
        imageUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required:
        - items
        - shippingAddress
        - deliveryMethod
        - paymentMethod
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        deliveryMethod:
          type: string
          enum: [standard, express]
        paymentMethod:
          type: string
          enum: [card]
    OrderItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          minimum: 1
        quantity:
          type: integer
          minimum: 1
    ShippingAddress:
      type: object
      required: [firstName, lastName, street, postalCode, city, country]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        street:
          type: string
        addressLine2:
          type: string
          nullable: true
        landmark:
          type: string
          nullable: true
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, totalPrice]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
    Order:
      type: object
      required:
        - id
        - userId
        - status
        - paymentStatus
        - deliveryMethod
        - paymentMethod
        - subtotal
        - shippingAmount
        - totalAmount
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
        paymentStatus:
          type: string
          enum: [PENDING, PAID, FAILED, CANCELLED]
        deliveryMethod:
          type: string
          enum: [standard, express]
        paymentMethod:
          type: string
          enum: [card]
        subtotal:
          type: number
          format: double
        shippingAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateOrderResponse:
      type: object
      required:
        - order
        - items
        - payment
      properties:
        order:
          $ref: '#/components/schemas/Order'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payment:
          $ref: '#/components/schemas/Payment'
    Payment:
      type: object
      required: [id, orderId, provider, amount, currency, status]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        provider:
          type: string
          enum: [STRIPE]
        amount:
          type: number
          format: double
        currency:
          type: string
          enum: [EUR]
        status:
          type: string
          enum: [PENDING, PAID, FAILED, CANCELLED]
        checkoutUrl:
          type: string
          format: uri
          nullable: true
...
    Store:
      type: object
      required:
        - id
        - location_id
        - name
        - address
        - timezone
      properties:
        id:
          type: integer
        location_id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        timezone:
          type: string
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Invalid request